\name{beta_emp}
\alias{beta_emp}
\docType{data}
\title{
Simulated methylation 450K array data with related
}
\description{
This dataset is a list containing two matrices, one of
which is methylation 450K array data of 3000 CpG sites on 50 samples, the other is
methylation 450K array data of 3000 matched CpG sites on three immune cell types.
The first dataset is generated by simulation. It originally has 459226 features and 50 samples.We reduce it to 3000 CpGs by random selection.
}
\usage{data("beta_emp")}
\format{
  The format is:
List of 2
 $ Y.raw: num [1:3000, 1:50] 0.7661 0.0968 0.8882 0.0286 0.6956 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:3000] "cg08752431" "cg14555682" "cg23086843" "cg20308511" ...
  .. ..$ : NULL
 $ ref.m: num [1:3000, 1:3] 0.7712 0.0996 0.9065 0.037 0.7242 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:3000] "cg08752431" "cg14555682" "cg23086843" "cg20308511" ...
  .. ..$ : chr [1:3] "CD4T" "CD8T" "BCell"
}
\examples{
data(beta_emp)
}
\keyword{datasets}
