% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ee_asset.R
\name{build_ee_asset}
\alias{build_ee_asset}
\title{build ee asset}
\usage{
build_ee_asset(DT, out, projection, id = "id", coords = c("X", "Y"),
  extra = NULL, overwrite = FALSE)
}
\arguments{
\item{out}{pathname of output shapefile folder. Includes location of the output folder but does not matter if that folder exists. The last portion of the pathname will be used for the layer name e.g.: 'path/to/caribou' layer name = 'caribou'}
}
\description{
Build an Earth Engine asset from a data.table of locs.
}
