% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ee_asset.R
\name{build_pol_asset}
\alias{build_pol_asset}
\title{build pol asset}
\usage{
build_pol_asset(pol, out, overwrite = FALSE)
}
\arguments{
\item{pol}{polygon}

\item{out}{pathname of output shapefile folder. Includes location of the output folder but does not matter if that folder exists. The last portion of the pathname will be used for the layer name e.g.: 'path/to/caribou' layer name = 'caribou'}

\item{overwrite}{boolean passed to option \code{overwrite_layer} in \code{writeOGR}}
}
\description{
build pol asset
}
