% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseScores.R
\name{compile_busco_summaries}
\alias{compile_busco_summaries}
\title{Get Busco scores from a series of summary files}
\usage{
result_df <- compile_busco_summaries(folderpath)
}
\arguments{
\item{folderpath}{The path to a root directory that contains directories that each represent the outputs from a busco run. Note that the name of each directory will be used as a sample ID}
}
\value{
Returns the busco scores in a dataframe along with the sample ID determined by the directory name
}
\description{
Fetch transcriptome completeness assessment from BUSCO outputs
}
\examples{
all_summaries <- compile_busco_summaries(folder_path)}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{score}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
