% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteCoverage.R
\name{CompleteCoverage}
\alias{CompleteCoverage}
\title{Scrutinize missing data patterns in your sequence data}
\usage{
CompleteCoverage(tsv)
}
\arguments{
\item{tsv}{A data frame of missing data coverage such as the generated “parsed_busco_results.tsv”}
}
\value{
Returns the names of taxa which have no missing data
}
\description{
Assess which taxa have no missing data
}
\examples{
CompleteCoverage(tsv)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{missing}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
