% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseScores.R
\name{csv2Fasta}
\alias{csv2Fasta}
\title{Convert a CSV or dataframe with sequences in a column into a fasta}
\usage{
csv2Fasta(input_csv, output_fasta="mysequences.fasta", external=TRUE,seqID="Genus",seq="sequence") 
}
\arguments{
\item{input_csv}{The path to a csv file or the name of a dataframe in memory from which to extract the sequences and header}
\item{output_fasta}{The name for your new fasta file}
\item{external}{Whether or not input_csv is a path to an external file. Default is TRUE. Use FALSE if you wish to run this function on a dataframe}
\item{seqID}{The name of the column that contains the names to use as headers for your fasta}
\item{seq}{The name of the column that contains the sequences}
}
\value{
Returns a fasta in the working directory
}
\description{
Convert a CSV or dataframe with sequences in a column into a fasta file}
\examples{
all_summaries <- compile_busco_summaries(folder_path)

csv2Fasta(input_csv="Path/to/your/file.fasta", output_fasta="mysequences.fasta", external=TRUE,seqID="Genus",seq="sequence")

#or if internal and you need to filter by an identifier and combine a bunch of things to make a header (Genus, species, accession in this example)

dataframe %>%
  	filter(Class== "targetTaxonomicRank") %>%
    mutate(fasta_id = paste0(">", genus, "_", species, "_", accession))
csv2Fasta(input_csv=dataframe, output_fasta="mysequences.fasta", external=FALSE,seqID="fasta_id",seq="sequence")

}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{utility}
\keyword{csv}

