% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_id.R
\name{dyad_id}
\alias{dyad_id}
\title{Dyad ID}
\usage{
dyad_id(DT = NULL, focal = "id", neighbour = NULL)
}
\arguments{
\item{DT}{data.table.}

\item{focal}{focal individual.}

\item{neighbour}{neighbour, optional - see Details.}
}
\description{
Build unique ID for dyads of focal individuals (and neighbours).
}
\details{
This function builds a dyad ID: individual A -> B and individual B -> A.

If neighbour is not provided, all potential dyads are returned as a "dyad dictionary".

Alternatively, when observed neighbours are provided, a "dyadID" column is added to the data.table with the individuals
}
