% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntrezDownload.R
\name{EntrezDownload}
\alias{EntrezDownload}
\title{Download Public Sequence Data}
\usage{
EntrezDownload(txid, fasta_dir, minimumSeq = 350, maximumSeq = NULL)
}
\arguments{
\item{txid}{NCBI taxon ID. See \url{https://ncbi.nlm.nih.gov} and manual}

\item{fasta_dir}{Place to store fasta sequences. Must end with */fasta/}

\item{minimumSeq}{Minimum number of sequences a species must have to bother downloading
defaults to 350}

\item{maximumSeq}{Maximum number of sequences to download per species.
Useful for testing scripts and troubleshooting TOAST/BUSCO}
}
\value{
Returns graphic of missing data based on defined threshold
}
\description{
Utility for downloading sequence data from NCBI for a user specified group of organisms
}
\examples{
SequenceDownload(txid = 9721, fasta_dir = "path/to/fasta/", minimumSeq = 350)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{entrez}
\keyword{fasta}
\keyword{missing}
\keyword{ncbi}
\keyword{phylogeny}
\keyword{sequence}
\keyword{taxa}
\keyword{taxonomy}
\keyword{toast}
\keyword{transcript}
