% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBuscos.R
\name{extractBuscos}
\alias{extractBuscos}
\title{Utility for extracting BUSCO sequences into individual fasta files from multiple directories.}
\usage{
extractBuscos(
  tsvLocations,
  fastaLocations,
  ed,
  SampleIDs,
  complete = TRUE,
  fragmented = TRUE,
  threshold = 300,
  duplicated = TRUE
)
}
\arguments{
\item{tsvLocations}{the paths that contain the tsv files from a busco analysis.}

\item{fastaLocations}{the paths to the original fasta sequence files}

\item{ed}{extracted directory where extracted sequences will be written}

\item{complete}{whether or not to include complete sequences}

\item{fragmented}{whether or not to include fragmented sequences}

\item{threshold}{minimum number of base pairs required for a fragmented sequence to be extracted.Not used when fragmented = false}

\item{duplicated}{whether or not to include duplicated sequences}

\item{sampleIDs}{name of sequences}
}
\value{
This function uses the output of a busco analysis and specified fasta 
        file to extract  busco sequences and write these into fasta files. 
        Fasta files are written into the directory specified by the parameter ed.
        Each sequence will be named based on the seqID specified
}
\description{
Utility for extracting BUSCO sequences into individual fasta files from multiple directories.
}
\examples{
extractBuscos(tsvLocations=c("path/to/first/tsvfile","path/to/second/tsvfile",...), fasta=c("path/to/first/fastafile","path/to/second/fastafile",...),ed="path/to/extracted/", sampleIDs=c("Genus_species","Othergenus_otherspecies",...),complete=TRUE, fragmented=TRUE, threshold=300, duplicated=TRUE)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}

Phillip Souza, \email{psouza1@uncc.edu}
}
\keyword{busco}
\keyword{extraction}
\keyword{fasta}
\keyword{toast}
\keyword{utility}
