% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBuscoSeqsFrag.R
\name{extractBuscoSeqsFrag}
\alias{extractBuscoSeqsFrag}
\title{Utility for extracting fragmented BUSCO sequences into individual fasta files}
\usage{
extractBuscoSeqsFrag(
  tsvtable,
  fasta,
  ed,
  seqID = "Genus_species",
  threshold = 300
)
}
\arguments{
\item{tsvtable}{the resulting tsv table from a busco analysis}

\item{fasta}{original fasta sequence file}

\item{ed}{extracted directory where extracted sequences will be written}

\item{seqID}{name of sequence, defaults to Genus_species}

\item{threshold}{minimum number of base pairs required for a fragmented sequence to be extracted. Defaults to 300}
}
\value{
This function uses the output of a busco analysis and specified fasta 
        file to extract fragmented busco sequences and write these into fasta files. 
        Fasta files are written into the directory specified by the parameter ed.
        Each sequence will be named based on the seqID specified
}
\description{
Utility for extracting fragmented BUSCO sequences into individual fasta files
}
\examples{
extractBuscoSeqsFrag(tsvtable=tsvtable, fasta=fasta,ed="path/to/extracted/", seqID="Genus_species",threshold=300)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}

Phillip Souza, \email{psouza1@uncc.edu}
}
\keyword{busco}
\keyword{extraction}
\keyword{fasta}
\keyword{toast}
\keyword{utility}
