% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseScores.R
\name{extractSequences}
\alias{extractSequences}
\title{Utility function for extracting sequences from HMMer outputs}
\usage{
extractSequences <- function(input_csv, type = "AA", external=TRUE) 
}
\arguments{
\item{input_csv}{The path to a csv file or the name of a dataframe in memory from which to extract the sequences and header}
\item{type}{AA for amino acid or DNA for nucleotide}
\item{external}{Whether or not input_csv is a path to an external file. Default is TRUE. Use FALSE if you wish to run this function on a dataframe}
}
\value{
Returns a fasta in the working directory
}
\description{
This function uses a table of sequence hits and target transcripts parsed from HMMer outputs to extract sequences }
\examples{
extractSequences <- function(input_csv="filepath/yourcsv.csv", type = "AA", external=TRUE) 

#or if internal and you have this in a dataframe already

extractSequences <- function(input_csv=results, type = "AA", external=FALSE) 
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{utility}
\keyword{csv}

