% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaIntoPhy.R
\name{FastaIntoPhy}
\alias{FastaIntoPhy}
\title{Utility for turning fasta file into phylip file}
\usage{
FastaIntoPhy(
  fastafile = "myfastadata",
  filename = "file_name",
  externalfile = TRUE
)
}
\arguments{
\item{fastafile}{defaults to myfastadata}

\item{filename}{defaults to file_name}

\item{externalfile}{defaults to TRUE}
}
\value{
This function uses an aligned fasta file that is either in your working directory 
        externalfile=TRUE or that has been read into memory, externalfile=false. 
        This then converts it to a phylip format for iqtree using the user specifed filename
}
\description{
Utility for turning fasta file into phylip file
}
\examples{
FastaIntoPhy(fastafile=myfastadata, filename="file_name",externalfile=TRUE)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{conversion}
\keyword{fasta}
\keyword{phylip}
\keyword{phylogeny}
\keyword{toast}
\keyword{utility}
