% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaIntoPhy.R
\name{geneTreeFilter}
\alias{geneTreeFilter}
\title{Utility for filtering genetrees based on ingroup monophyly or coverage}
\usage{
geneTreeFilter(
ingroup=ingroupTaxa, 
outgroup=outgroupTaxa, 
threshold=1, 
tree_file_dir="path/to/files", 
tree_file_ext=".tre"
)
}
\arguments{
\item{ingroup}{input vector of ingroups}

\item{outgroup}{input vector of vector of outgroups}

\item{threshold}{the minimum number of ingroup sequences, defaults to 1}

\item{tree_file_dir}{string of the directory path for tree files}

\item{tree_file_ext}{string extension of tree files defaults to ".contree"}
}
\value{
This function uses a user supplied vector of a target ingroup and outgroup taxa along with a path 
         to a directory containing newick format genetrees and their extension (e.g. ".tre", ".contree", etc) to assess sufficient taxon coverage 
         within a specific ingroup using a user defined threshold, and whether the ingroup of interest is resolved as monophyletic.
         The function returns a dataframe that gives each genetree name, the number of ingroup taxa present, and whether
         these pass inspection. Trees that do not pass due to low coverage, lack of outgroups, or non-monophyly are indicated by the term "omit".
}
\description{
Utility for filtering genetrees based on ingroup monophyly or coverage
}
\examples{
ingroupTaxa<-c("TaxonA", "TaxonB", "TaxonC", "TaxonD", "TaxonE")
outgroupTaxa<-c("TaxonZ", "TaxonY", "TaxonX")
geneTreeFilter(ingroup=ingroupTaxa, outgroup=outgroupTaxa, threshold=1, tree_file_dir="path/to/files", tree_file_ext=".tre")
}
\author{
Derrek Pope, \email{dpope8@uncc.edu}
}
\keyword{toast}
\keyword{phylogeny}
\keyword{genetree}
\keyword{filtration}
\keyword{missing data}
\keyword{utility}
