% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MafftOrientAlign.R
\name{MafftOrientAlign}
\alias{MafftOrientAlign}
\title{Align BUSCO Orthologs}
\usage{
MafftOrientAlign(extract_dir, mafft_dir, threads = 1)
}
\arguments{
\item{extract_dir}{Directory containing unaligned/unorientated fasta files}

\item{mafft_dir}{Directory to write orientated and aligned fasta files}

\item{threads}{Number of cpu threads to use}
}
\description{
This function allows you align each ortholog between all species
}
\examples{
MafftOrientAlign(extract_dir = "path/to/extracted/", mafft_dir = "/path/to/mafft/", threads = 1)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{align}
\keyword{alignment}
\keyword{busco}
\keyword{fasta}
\keyword{mafft}
\keyword{missing}
\keyword{orient}
\keyword{orientate}
\keyword{ortholog}
\keyword{phy}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
