% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingDataTable.R
\name{MissingDataTable}
\alias{MissingDataTable}
\title{Utility for Assessing Missing Data in a Set of FASTA Files}
\usage{
MissingDataTable(aligned_dir)
}
\arguments{
\item{aligned_dir}{Directory where fasta alignments can be found}
}
\value{
Reads through the full table found in folders ./busco_results/run_Busco* and
        returns a dataframe of the results
}
\description{
Parse through a directory of FASTA files to assess missing data patterns between taxa and loci
}
\examples{
missing_data_df <- MissingDataTable(aligned_dir)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{missing}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
