% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NexIntoFas.R
\name{NexIntoFas}
\alias{NexIntoFas}
\title{Utility for turning TOAST superalign.txt into Nexus file}
\usage{
NexIntoFas(nexusfile= myphylipdata, filename="file_name",externalfile=TRUE)
}
\arguments{
\item{filename}{defaults to superalign.txt assuming it is in the working directory. A full path can be specified}
}
\value{
This function uses an aligned phylip file that is either in your working directory 
externalfile=TRUE or that has been read into memory, externalfile=false. 
This then converts it to a fasta format writing to file with user specified filename
}
\description{
Utility for turning a nexus file into fasta file in TOAST
}
\examples{
NexIntoFas(nexusfile= myphylipdata, filename="file_name",externalfile=TRUE)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{nexus}
\keyword{toast}
\keyword{fasta}
\keyword{alignment}
\keyword{phylogeny}
\keyword{sequence}