% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseScores.R
\name{ParseScores}
\alias{ParseScores}
\title{Get Busco scores from summary file}
\usage{
result_df <- parse_busco_output(file_path, input_id)
}
\arguments{
\item{file_path}{The path to the output short summary from busco “parsed_busco_results.tsv”}
\item{input_id}{A usr defined name for the sample in case you want to plot this with a phylogeny or track with multiple samples. Examples include "Genus_species", "Accession", "strain", etc}
}
\value{
Returns the busco scores in a dataframe along with the sample ID
}
\description{
Fetch transcriptome completeness assessment from BUSCO output
}
\examples{
result_df <- parse_busco_output(file_path, input_id)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{score}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
