% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyIntoFasta.R
\name{PhyIntoFasta}
\alias{PhyIntoFasta}
\title{Utility for turning phylip file into fasta file}
\usage{
PhyIntoFasta(
  phylipfile = "myphylipdata",
  filename = "file_name",
  externalfile = TRUE
)
}
\arguments{
\item{phylipfile}{defaults to myphylipdata}

\item{filename}{defaults to file_name}

\item{externalfile}{defaults to TRUE}
}
\value{
This function uses an aligned phylip file that is either in your working directory 
        externalfile=TRUE or that has been read into memory, externalfile=false. 
        This then converts it to a fasta format writing to file with user specified filename
}
\description{
Utility for turning phylip file into fasta file
}
\examples{
PhyIntoFasta(fastafile= myphylipdata, filename="file_name",externalfile=TRUE)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{conversion}
\keyword{fasta}
\keyword{phylip}
\keyword{phylogeny}
\keyword{toast}
\keyword{utility}
