% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyIntoNex.R
\name{PhyIntoNex}
\alias{PhyIntoNex}
\title{Utility for turning TOAST superalign.txt into Nexus file}
\usage{
PhyIntoNex(filename = "superalign.txt")
}
\arguments{
\item{filename}{defaults to superalign.txt assuming it is in the working directory. A full path can be specified}
}
\value{
exports an alignment in classic phyllip format into a nexus file. 
This is designed for the superalign file generated by toast, 
but will work on any alignment file formatted the same way. 
}
\description{
Utility for turning TOAST superalign.txt into Nexus file
}
\examples{
PhyIntoNex(filename = "superalign.txt")
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{nexus}
\keyword{toast}
\keyword{nexus}
\keyword{alignment}
\keyword{phylogeny}
\keyword{sequence}