% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyIntoNex.R
\name{PhyIntoNex}
\alias{PhyIntoNex}
\title{Utility for turning TOAST superalign.txt into Nexus file}
\usage{
phyloMissmap(missing, tree)
}
\arguments{
\item{missing}{missing data object from TOAST}
\item{tree}{Tree in newick format}
\item{...}{Optional graphic parameters for formatting}
}
\value{
Plots a missmap plot of your data sorted by the tree tips
}
\description{
Plot function to create a tree sorted missmap plot
}
\examples{
phyloMissmap(missing, tree)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{nexus}
\keyword{toast}
\keyword{nexus}
\keyword{alignment}
\keyword{phylogeny}
\keyword{sequence}
\keyword{missing data}