% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBuscoSummaries.R
\name{PlotBuscoSummaries}
\alias{PlotBuscoSummaries}
\title{plot Busco scores from the compile_busco_summaries output}
\usage{
result_df <- compile_busco_summaries(folderpath)
PlotBuscoSummaries(result_df)
}
\arguments{
\item{result_df}{The output from the compile busco summaries function or dataframe in identical format}
\item{CSC}{Color choice for "Complete_Single_Copy" = CSC,  # Default=Light Blue}
\item{CDC}{Color choice for CDC "Complete_Duplicated" = CDC,   # Default=Steel Blue}
\item{Frag}{Color choice for Frag "Fragmented" = Frag,            # Default=Medium Purple}
\item{Miss}{Color choice for Miss "Missing" = Miss                # Default=Peach Puff}
}
\value{
Returns a plot
}
\description{
Graph transcriptome completeness assessment from BUSCO outputs
}
\examples{
all_summaries <- compile_busco_summaries(folder_path)
PlotBuscoSummaries(all_summaries)
}
\author{
Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{score}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
