% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunBusco.R
\name{RunBusco}
\alias{RunBusco}
\title{Run BUSCO to Assign Ortholog Ids}
\usage{
RunBusco(fasta_dir, toast_dir, path_to_run_busco.py, path_to_orthoDB,
  threads = 1)
}
\arguments{
\item{fasta_dir}{directory of fastas; Genus_species.fasta}

\item{toast_dir}{working directory where busco runs will be placed}

\item{path_to_run_busco.py}{full path where the python script can be found}

\item{threads}{number of threads to use, defaults to 1}

\item{path_to_orthDB}{directory where orthoDB can be found}
}
\value{
Runs BUSCO on all fasta sequences in fasta_dir
}
\description{
This function allows you to run BUSCO and harvest orthologs.
       Note that this function is limited to linux users as BUSCO will only run on linux
}
\examples{
RunBusco(fasta_dir = "path/to/fasta/", toast_dir = "/path/to/toast/dir/",
          path_to_run_busco.py = "path/to/run/busco.py/", path_to_orthoDB = "path/to/orthoDB")
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{missing}
\keyword{odb9}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
