% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_length.R
\name{step_length}
\alias{step_length}
\title{Step Length}
\usage{
step_length(DT, coords = c("EASTING", "NORTHING"), id = "id",
  time = "datetime", type = "lag", moverate = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector, length 2, coordinate column names. UTM required.}

\item{time}{character time column name.}

\item{type}{default: lag. alternative: lead.}

\item{moverate}{calculate movement rate? stepLength / dif time, unit hours.}
}
\description{
Calculate step length with data.table
}
\examples{

}
