% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_sample.R
\name{strat_sample}
\alias{strat_sample}
\title{Stratified polygon sampling}
\usage{
strat_sample(x, n, col, returnDT = TRUE)
}
\arguments{
\item{returnDT}{}
}
\value{
Either a \code{sf} object or a \code{data.table} with a \code{sfc} (simple feature column).
}
\description{
For each unique value in 'col', sample 'n' points and optionally return as a \code{data.table}.
}
