% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperAlign.R
\name{SuperAlign}
\alias{SuperAlign}
\title{Utility for Creating a concatenated alignment from Fasta Alignments}
\usage{
SuperAlign(aligned_dir, missing_df)
}
\arguments{
\item{aligned_dir}{Directory where fasta alignments can be found}
\item{missing_df}{Missing data object made by the missing function}
\item{filename}{String for what you want to name your output, default is "superalign.phy"}
}
\value{
Reads through the full table found in folders ./busco_results/run_Busco* and
        returns a dataframe of the results
}
\description{
Parse through a directory of FASTA files to combine into one alignment
}
\examples{
partition_df <- SuperAlign(aligned_dir, missing_df, filename="myalignment.phy")
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{missing}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
