% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOAST_coprimary.R
\name{three_outcomes_bin_count_coprimary}
\alias{three_outcomes_bin_count_coprimary}
\title{Three outcome approach decision making for continous + count co-primary endpoints}
\usage{
three_outcomes_bin_count_coprimary(
  n1_prev,
  n2_prev,
  end_1_p1,
  end_1_p2,
  end_2_shape1,
  end_2_rate1,
  end_2_shape2,
  end_2_rate2,
  nrep,
  NREP,
  sample_size,
  independent = T,
  correlation = NULL,
  strategy = "conservative",
  LRV_1,
  LRV_2,
  TV_1,
  TV_2,
  FGR = 0.2,
  FSR = 0.1,
  seed
)
}
\arguments{
\item{n1_prev}{Numeric. Sample size from control group in the previous studies}

\item{n2_prev}{Numeric. Sample size from treatment group in the previous studies}

\item{end_1_p1}{Numeric. Response rate from control group in the previous studies for endpoints one}

\item{end_1_p2}{Numeric. Response rate from treatment group in the previous studies for endpoints one}

\item{end_2_shape1}{Numeric. Shape parameter to generate Poisson rate for endpoints two from control group in the previous studies}

\item{end_2_rate1}{Numeric. Rate parameter to generate Poisson rate for endpoints two from control group in the previous studies}

\item{end_2_shape2}{Numeric. Shape parameter to generate Poisson rate for endpoints two from treatment group in the previous studies}

\item{end_2_rate2}{Numeric. Rate parameter to generate Poisson rate for endpoints two from treatment group in the previous studies}

\item{nrep}{Numeric. Replicates to evaluate the decision once}

\item{NREP}{Numeric. Replicates to calculate the decision proportions}

\item{sample_size}{Numeric. Target Sample size in the new studies}

\item{independent}{Logical. TRUE if two endpoints are independent or FALSE if they are correlated}

\item{correlation}{Numeric. correlation strength between two endpoints if they are correlated}

\item{strategy}{Character. Two values: conservative or aggressive for decision criteria}

\item{LRV_1}{Numeric. Lower reference value (treatment difference) for endpoint one}

\item{LRV_2}{Numeric. Lower reference value (treatment difference) for endpoint two}

\item{TV_1}{Numeric. Target value (treatment difference) for endpoint one}

\item{TV_2}{Numeric. Target value (treatment difference) for endpoint two}

\item{FGR}{Numeric. False go risk level}

\item{FSR}{Numeric. False stop risk level}

\item{seed}{Numeric. Random seed}
}
\description{
Three outcome approach decision making for continous + count co-primary endpoints
}
\examples{
library(TOAST)
## Example code for two binary primary endpoint
bin_count_examples_ind <- three_outcomes_bin_count_coprimary(n1_prev = 50, n2_prev = 50,
                                              end_1_p1 = 0.3, end_1_p2 = 0.5,
                                              end_2_shape1 = 3, end_2_rate1 = 5,
                                              end_2_shape2 = 5, end_2_rate2 = 6,
                        nrep = 300, NREP = 200, sample_size = 100, strategy = "conservative",
                        LRV_1 = 0.1, LRV_2 = 0.1, independent = T,
                        TV_1 = 0.2, TV_2 = 0.233, FGR = 0.2, FSR = 0.1, seed = 2024)
bin_count_examples_cor <- three_outcomes_bin_count_coprimary(n1_prev = 50, n2_prev = 50,
                                              end_1_p1 = 0.3, end_1_p2 = 0.5,
                                              end_2_shape1 = 3, end_2_rate1 = 5,
                                              end_2_shape2 = 5, end_2_rate2 = 6,
                        nrep = 300, NREP = 200, sample_size = 100, strategy = "conservative",
                        LRV_1 = 0.1, LRV_2 = 0.1, independent = F, correlation = 0.3,
                        TV_1 = 0.2, TV_2 = 0.233, FGR = 0.2, FSR = 0.1, seed = 2024)

}
