% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOAST.R
\name{three_outcomes_binary}
\alias{three_outcomes_binary}
\title{TOAST for binary type of endpoints}
\usage{
three_outcomes_binary(
  n1_prev,
  n2_prev,
  p_1,
  p_2,
  Method = NULL,
  nrep,
  NREP,
  sample_size,
  LRV,
  TV,
  FGR = 0.2,
  FSR = 0.1,
  seed
)
}
\arguments{
\item{n1_prev}{Numeric. Sample size from control group in the previous studies}

\item{n2_prev}{Numeric. Sample size from treatment group in the previous studies}

\item{p_1}{Numeric. Response rate from control group in the previous studies}

\item{p_2}{Numeric. Response rate from treatment group in the previous studies}

\item{Method}{Character. Method chosen for confidence interval calculation - empirical or bootstrap}

\item{nrep}{Numeric. Replicates to evaluate the decision once}

\item{NREP}{Numeric. Replicates to calculate the decision proportions}

\item{sample_size}{Numeric. Target Sample size in the new studies}

\item{LRV}{Numeric. Lower reference value (treatment difference) defined prior to the studies}

\item{TV}{Numeric. Target value (treatment difference) defined prior to the studies}

\item{FGR}{Numeric. False go risk level}

\item{FSR}{Numeric. False stop risk level}

\item{seed}{Numeric. Simulation seed for replication purpose}
}
\value{
A list containing the population parameters/ TPP values/ number of simulations/ decision results
}
\description{
TOAST for binary type of endpoints
}
\examples{
library(TOAST)
## Example code for one binary primary endpoint
binary_examples <- three_outcomes_binary(n1_prev = 60, n2_prev = 60, p_1 = 0.1, p_2 = 0.2,
   nrep = 500, NREP = 1000, sample_size = 75,
   LRV = 0.093, TV = 0.1, Method = "empirical",
   FGR = 0.2, FSR = 0.1, seed = 2024)
}
