% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOAST_coprimary.R
\name{three_outcomes_binary_coprim}
\alias{three_outcomes_binary_coprim}
\title{Three outcome approach decision making for binary type of endpoints
The limits are determined by the CLT only}
\usage{
three_outcomes_binary_coprim(
  n1_prev,
  n2_prev,
  end_1_p1,
  end_2_p1,
  end_1_p2,
  end_2_p2,
  nrep,
  NREP,
  sample_size,
  independent = T,
  correlation = NULL,
  LRV_1,
  LRV_2,
  TV_1,
  TV_2,
  seed
)
}
\arguments{
\item{n1_prev}{Numeric. Sample size from control group in the previous studies}

\item{n2_prev}{Numeric. Sample size from treatment group in the previous studies}

\item{end_1_p1}{Numeric. Response rate from control group in the previous studies for endpoints one}

\item{end_2_p1}{Numeric. Response rate from control group in the previous studies for endpoints two}

\item{end_1_p2}{Numeric. Response rate from treatment group in the previous studies for endpoints one}

\item{end_2_p2}{Numeric. Response rate from treatment group in the previous studies for endpoints two}

\item{nrep}{Numeric. Replicates to evaluate the decision once}

\item{NREP}{Numeric. Replicates to calculate the decision proportions}

\item{sample_size}{Numeric. Target Sample size in the new studies}

\item{independent}{Logical. TRUE if two endpoints are independent or FALSE if they are correlated}

\item{correlation}{Numeric. correlation strength between two endpoints if they are correlated}

\item{LRV_1}{Numeric. Lower reference value (treatment difference) for endpoint one}

\item{LRV_2}{Numeric. Lower reference value (treatment difference) for endpoint two}

\item{TV_1}{Numeric. Target value (treatment difference) for endpoint one}

\item{TV_2}{Numeric. Target value (treatment difference) for endpoint two}

\item{seed}{Numeric. Random seed}
}
\description{
Three outcome approach decision making for binary type of endpoints
The limits are determined by the CLT only
}
