% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOAST.R
\name{three_outcomes_cont}
\alias{three_outcomes_cont}
\title{TOAST for continuous type of endpoints}
\usage{
three_outcomes_cont(
  n1_prev,
  n2_prev,
  mean_1,
  mean_2,
  std_1,
  std_2,
  Method = NULL,
  nrep,
  NREP,
  sample_size,
  LRV,
  TV,
  FGR = 0.2,
  FSR = 0.1,
  seed
)
}
\arguments{
\item{n1_prev}{Numeric. Sample size from control group in the previous studies}

\item{n2_prev}{Numeric. Sample size from treatment group in the previous studies}

\item{mean_1}{Numeric. Sample mean from control group in the previous studies}

\item{mean_2}{Numeric. Sample mean from treatment group in the previous studies}

\item{std_1}{Numeric. Std for the mean parameter from control group in the previous studies}

\item{std_2}{Numeric. Std for the mean parameter from treatment group in the previous studies}

\item{Method}{Character. Method chosen for confidence interval calculation - empirical or bootstrap}

\item{nrep}{Numeric. Replicates to evaluate the decision once}

\item{NREP}{Numeric. Replicates to calculate the decision proportions}

\item{sample_size}{Numeric. Target Sample size in the new studies}

\item{LRV}{Numeric. Lower reference value (treatment difference) defined prior to the studies}

\item{TV}{Numeric. Target value (treatment difference) defined prior to the studies}

\item{FGR}{Numeric. False go risk level}

\item{FSR}{Numeric. False stop risk level}

\item{seed}{Numeric. Random seed}
}
\value{
A list containing the population parameters/ TPP values/ number of simulations/ decision results
}
\description{
TOAST for continuous type of endpoints
}
\examples{
library(TOAST)
## Example code for one continuous primary endpoint
cont_examples <- three_outcomes_cont(n1_prev = 50, n2_prev = 50,
mean_1 = 1.2, mean_2 = 1.45, std_1 = 0.3, std_2 = 0.3,
Method = "empirical", nrep = 500, NREP = 1000, sample_size = 100,
FGR = 0.2, FSR = 0.1,
LRV = 0.2, TV = 0.25, seed = 2024)

}
