% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOAST.R
\name{three_outcomes_survival}
\alias{three_outcomes_survival}
\title{TOAST for time to event type of endpoints (survival data) using Exponential distribution}
\usage{
three_outcomes_survival(
  hazard_rate_1,
  HR,
  rate_1,
  rate_2,
  censoring_factor = 2,
  nrep,
  NREP,
  sample_size,
  LRV,
  TV,
  FGR = 0.2,
  FSR = 0.1,
  seed
)
}
\arguments{
\item{hazard_rate_1}{Numeric. 1 over Mean estimate for time to event from control group in the previous studies}

\item{HR}{Numeric. HR comparing the treatment group to the control group}

\item{rate_1}{Numeric. Rate parameter to control the failure rate (gamma distribution) from control group in the previous studies}

\item{rate_2}{Numeric. Shape parameter to control the failure rate (gamma distribution) from treatment group in the previous studies}

\item{censoring_factor}{Numeric. Censoring factor - the mean of censoring time would be the original mean times this factor}

\item{nrep}{Numeric. Replicates to evaluate the decision once}

\item{NREP}{Numeric. Replicates to calculate the decision proportions}

\item{sample_size}{Numeric. Target Sample size in the new studies}

\item{LRV}{Numeric. Lower reference value (treatment difference) defined prior to the studies}

\item{TV}{Numeric. Target value (treatment difference) defined prior to the studies}

\item{FGR}{Numeric. False go risk level}

\item{FSR}{Numeric. False stop risk level}

\item{seed}{Numeric. Simulation seed for replication purpose}
}
\value{
A list containing the population parameters/ TPP values/ number of simulations/ decision results
}
\description{
TOAST for time to event type of endpoints (survival data) using Exponential distribution
}
\examples{
library(TOAST)
## Example code for one time to event primary endpoint
three_outcomes_survival(hazard_rate_1 = 1/60, HR = 0.6,
   rate_1 = 1, rate_2 = 1, censoring_factor = 5,
   nrep = 500, NREP = 1000, sample_size = 100,
   FGR = 0.2, FSR = 0.1, LRV = 0.8, TV = 0.6, seed = 2024)
}
