% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdDataTable.R
\name{ThresholdDataTable}
\alias{ThresholdDataTable}
\title{Utility for retrieving species sequences givin a missing data threshold}
\usage{
ThresholdDataTable(missing_df, threshold)
}
\arguments{
\item{missing_df}{Directory where fasta alignments can be found}

\item{threshold}{Minimum number of sequences a sample/species must contain}
}
\value{
Reads through the full table found in folders ./busco_results/run_Busco* and
        returns a dataframe of the results
}
\description{
User defines a threshold and is returned a dataframe containing
       only the samples/species with at least that many sequences.
       This dataframe can be fed into ExtractBuscoSeqs() as the busco_table
       parameter
}
\examples{
threshold_df <- ThresholdDataTable(missing_df = mdf, threshold = 50)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{missing}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
