% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdExtract.R
\name{ThresholdExtract}
\alias{ThresholdExtract}
\title{Utility for retrieving species sequences givin a dataframe of missing data}
\usage{
ThresholdExtract(aligned_dir, missing_df, threshold_fasta_folder)
}
\arguments{
\item{aligned_dir}{Directory where fasta alignments can be found}

\item{missing_df}{Dataframe created by ParseBuscoResults() or MissingDataTable()}

\item{threshold_fasta_folder}{Folder to store extracted fasta files}
}
\value{
Reads through a dataframe of missing and extracts sequences from the aligned_dir
        and writes to threshold_fasta_folder. Fasta files must be realigned and this can be
        accomplished with MafftOrientAlign()
}
\description{
User defines a threshold and is returned a dataframe containing
       only the samples/species with at least that many sequences.
       This dataframe can be fed into ExtractBuscoSeqs() as the busco_table
       parameter
}
\examples{
ThresholdExtract(aligned_dir = ad, missing_df = threshold_df, threshold_fasta_folder = "/path/to/store/fastas/threshold100")
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{missing}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
