% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{TOAST_shiny}
\alias{TOAST_shiny}
\title{Run the TOAST Shiny Application}
\usage{
TOAST_shiny()
}
\description{
Run the TOAST Shiny Application
}
\examples{

\dontrun{
 ## Will Load an Interactive Session
shinyOutput <- TOAST_shiny()
}
}
