% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToastGoSlim.R
\name{ToastGoSlim}
\alias{ToastGoSlim}
\title{Assign Go_Slim IDs to BUSCO IDs}
\usage{
ToastGoSlim(orthogroup_info, obo, perspective = "BP")
}
\arguments{
\item{orthogroup_info}{assing the uncompressed ontology file found in orthoDB/info/*orthogroup_info.txt.gz}

\item{obo}{download desired ontology information (.obo file) from http://geneontology.org/docs/download-ontology/#go_obo_and_owl
these are consistently updated so make sure to grab the newest
we recommend the GO slim AGR subset (goslim_agr.obo) which can be obtained using the command
wget http://current.geneontology.org/ontology/subsets/goslim_agr.obo}

\item{perspective}{options are BP(biological process), MF(Molecular function), or CC(cellular component)}
}
\description{
this function takes the Gene Ontology information contained in the orthoDB and generalizes it
        into as as few GO terms as possible given three perspectives: Molecular function,
        Biological Process and Cellular Component, each returned as a dataframe
}
\examples{
BP <- ToastGoSlim(orthogroup_info = "path/to/orthoDB/info/*orthogroup_info.txt.gz", obo = "path/to/goslim_agr.obo", perspective = "BP")
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{busco}
\keyword{fasta}
\keyword{gene}
\keyword{go}
\keyword{missing}
\keyword{ontology}
\keyword{ortholog}
\keyword{phylogeny}
\keyword{sequence}
\keyword{slim}
\keyword{toast}
\keyword{transcript}
