% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeCoverage.R
\name{VisualizeCoverage}
\alias{VisualizeCoverage}
\title{Explore Global Patterns of Missing Data}
\usage{
VisualizeCoverage(tsv, threshold = 100)
}
\arguments{
\item{tsv}{A data frame of missing data coverage such as the generated “parsed_busco_results.tsv”}

\item{threshold}{Cutoff value for minimum number of sequences allowed per taxon}
}
\description{
Assess coarse missing data patterns using circlepack plots and barplots
}
\examples{
VisualizeCoverage(tsv, threshold)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{decisiveness}
\keyword{missing}
\keyword{phylogenomics}
\keyword{sequence}
\keyword{toast}
