% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePBR.R
\name{VisualizePBR}
\alias{VisualizePBR}
\title{Explore Global Patterns of Missing Data}
\usage{
VisualizePBR(busco_dir)
}
\arguments{
\item{tsv}{A data frame of missing data coverage such as the generated “parsed_busco_results.tsv”}

\item{threshold}{Cutoff value for minimum number of sequences allowed per taxon}
}
\description{
Returns a dataframe that can be used to generated graphics of missing data patterns
}
\examples{
VPBR <- isualizePBR(busco_dir = bd)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{decisiveness}
\keyword{missing}
\keyword{phylogenomics}
\keyword{sequence}
\keyword{toast}
