% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeTaxonomyInteractive.R
\name{VisualizeTaxonomyInteractive}
\alias{VisualizeTaxonomyInteractive}
\title{Explore Missing Data Interactively Based on a Hierarchical Structure}
\usage{
VisualizeTaxonomyInteractive(tsv, taxonomy, threshold)
}
\arguments{
\item{tsv}{a data frame of missing data coverage such as the generated “parsed_busco_results.tsv”}

\item{taxonomy}{user made data.frame of taxonomy levels following the format of cetacean_taxonomy}

\item{threshold}{cutoff value for minimum number of sequences allowed per taxon}
}
\description{
This function allows you generate circlepack plots to assess patterns of missing data at different hierarchical levels interactively
}
\examples{
VisualizeTaxonomyInteractive(tsv, taxonomy, threshold)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{barplot}
\keyword{data}
\keyword{missing}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
