% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeThreshold.R
\name{VisualizeThreshold}
\alias{VisualizeThreshold}
\title{Determine and Appropriate Missing Data Threshold}
\usage{
VisualizeThreshold(tsv, threshold)
}
\arguments{
\item{tsv}{a data frame of missing data coverage such as the generated “parsed_busco_results.tsv”}

\item{threshold}{cutoff value for minimum number of sequences allowed per taxon}
}
\value{
Returns graphic of missing data based on defined threshold
}
\description{
Graphically assess the impact of a missing data threshold on changes in the distribution of missing data
}
\examples{
VisualizeThreshold("parsed_busco_results.tsv", 1000)
}
\author{
Dustin J Wcisel, \email{djwcisel@ncsu.edu}

James Thomas Howard, \email{jthowar3@ncsu.edu}

Jeffrey A Yoder, \email{jayoder@ncsu.edu}

Alex Dornburg, \email{dornburgalex@gmail.com}
}
\keyword{DNA}
\keyword{missing}
\keyword{phylogeny}
\keyword{sequence}
\keyword{toast}
\keyword{transcript}
