% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_id.R
\name{dyad_id}
\alias{dyad_id}
\title{Dyad ID}
\usage{
dyad_id(DT = NULL, focal = "id", neighbour = NULL, both = FALSE)
}
\arguments{
\item{DT}{data.table.}

\item{focal}{focal individual.}

\item{neighbour}{neighbours, optional - see Details.}

\item{both}{if both is TRUE, returns duplicated dyadIDs with all focal/neighbour inverted and appended.}
}
\value{
If neighbour is not provided, all potential dyads are returned as a "dyad dictionary".

Alternatively, when observed neighbours are provided, a "dyadID" column is added to the data.table with the individuals
}
\description{
Build unique ID for dyads of focal individuals (and neighbours).
}
\details{
This function builds a dyad ID: focal A -> neighbour B (and focal B -> neighbour A) = dyad 1.
}
\examples{
# Load data.table
library(data.table)

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "toast"))

dyadDict <- dyad_id(DT, focal = 'ID', neighbour = NULL)
}
