% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOAST_coprimary.R
\name{three_outcomes_count_coprim}
\alias{three_outcomes_count_coprim}
\title{Three outcome approach decision making for count type of endpoints
The limits are determined by the CLT only}
\usage{
three_outcomes_count_coprim(
  end_1_shape1,
  end_1_rate1,
  end_2_shape1,
  end_2_rate1,
  end_1_shape2,
  end_1_rate2,
  end_2_shape2,
  end_2_rate2,
  nrep,
  NREP,
  sample_size,
  independent = T,
  correlation = NULL,
  LRV_1,
  LRV_2,
  TV_1,
  TV_2,
  seed
)
}
\arguments{
\item{nrep}{Numeric. Replicates to evaluate the decision once}

\item{NREP}{Numeric. Replicates to calculate the decision proportions}

\item{sample_size}{Numeric. Target Sample size in the new studies}

\item{independent}{Logical. TRUE if two endpoints are independent or FALSE if they are correlated}

\item{correlation}{Numeric. correlation strength between two endpoints if they are correlated}

\item{LRV_1}{Numeric. Lower reference value (treatment difference) for endpoint one}

\item{LRV_2}{Numeric. Lower reference value (treatment difference) for endpoint two}

\item{TV_1}{Numeric. Target value (treatment difference) for endpoint one}

\item{TV_2}{Numeric. Target value (treatment difference) for endpoint two}

\item{seed}{Numeric. Random seed}

\item{end_1_m1}{Numeric. Sample mean for endpoints one from control group in the previous studies}

\item{end_1_std1}{Numeric. Std for endpoints one from control group in the previous studies}

\item{end_2_m1}{Numeric. Sample mean for endpoints two from control group in the previous studies}

\item{end_2_std1}{Numeric. Std for endpoints two from control group in the previous studies}

\item{end_1_m2}{Numeric. Sample mean for endpoints one from treatment group in the previous studies}

\item{end_1_std2}{Numeric. Std for endpoints one from treatment group in the previous studies}

\item{end_2_m2}{Numeric. Sample mean for endpoints two from treatment group in the previous studies}

\item{end_2_std2}{Numeric. Std for endpoints two from treatment group in the previous studies}
}
\description{
Three outcome approach decision making for count type of endpoints
The limits are determined by the CLT only
}
